refactor(frontend): relocate resolveBadgeBucket to satisfactionLookup
