+++ frontend/src/utils/satisfactionLookup.ts
@@ -12,9 +12,10 @@
  * which suppressed these legitimate API-provided strings.
  */
 
-import type { PerRequestStatus, SatisfactionEntry } from '../types/satisfaction'
+import type { PerRequestStatus, RequestBucket, SatisfactionEntry } from '../types/satisfaction'
 import { isAgePreferenceSatisfied } from './agePreferenceSatisfaction'
 import { formatGradeOrdinal } from './gradeUtils'
+import { safeSourceFromField } from './sourceFromField'
 import type { EnhancedBunkRequest } from '../hooks/camper/useAllBunkRequests'
 
 export function buildSatisfactionLookup(
@@ -147,4 +148,33 @@ export function evaluateRequest({
   }
 
   return { status: 'unknown' }
+}
+
+/**
+ * Resolve the P/S age-preference badge state for a request row (#1172).
+ *
+ * The centralized aggregator (`CamperSatisfaction.per_request[i].bucket`) is
+ * the source of truth when present — that's what #1158/#1159 consolidated.
+ * When the aggregator is unavailable (`/api/satisfaction` 500, empty response,
+ * loading state), bucket is `undefined` and the badge silently disappears
+ * unless we fall back to the row's own `source_field`/`source` — which is
+ * what drove the badge pre-#1158 and cannot fail.
+ *
+ * Centralized bucket wins; row-level fields are ONLY consulted when the
+ * centralized map has no entry for this row.
+ */
+export function resolveBadgeBucket(
+  bucket: RequestBucket | undefined,
+  req: { source_field?: string | null; source?: string | null; request_type?: string | null }
+): { isMaterialAgePref: boolean; isStaffBadge: boolean } {
+  if (bucket === 'material_parent') return { isMaterialAgePref: true, isStaffBadge: false }
+  if (bucket === 'staff') return { isMaterialAgePref: false, isStaffBadge: true }
+  if (bucket === undefined) {
+    const derived = safeSourceFromField(req.source_field)
+    return {
+      isMaterialAgePref: req.request_type === 'age_preference' && req.source_field === 'bunk_with',
+      isStaffBadge: derived !== null ? derived === 'staff' : req.source === 'staff',
+    }
+  }
+  return { isMaterialAgePref: false, isStaffBadge: false }
 }
